+++ core/src/androidMain/kotlin/nl/adaptivity/xmlutil/XmlStreaming.android.kt
@@ -25,16 +25,26 @@ import nl.adaptivity.xmlutil.XmlStreaming.deSerialize
 import nl.adaptivity.xmlutil.core.KtXmlReader
 import nl.adaptivity.xmlutil.core.KtXmlWriter
 import nl.adaptivity.xmlutil.core.impl.XmlStreamingJavaCommon
+import nl.adaptivity.xmlutil.core.impl.multiplatform.Writer
 import nl.adaptivity.xmlutil.util.SerializationProvider
-import java.io.*
+import java.io.InputStream
+import java.io.OutputStream
+import java.io.Reader
+import java.io.StringReader
 import java.util.*
 import javax.xml.transform.Result
 import javax.xml.transform.Source
 import kotlin.reflect.KClass
 
 
-public actual object XmlStreaming : XmlStreamingJavaCommon() {
+@Deprecated("Don't use directly", ReplaceWith("xmlStreaming",
+    "nl.adaptivity.xmlutil.xmlStreaming",
+    "nl.adaptivity.xmlutil.newWriter",
+    "nl.adaptivity.xmlutil.newGenericWriter",
+))
+public actual object XmlStreaming : XmlStreamingJavaCommon(), IXmlStreaming {
 
+    @Deprecated("This functionality uses service loaders and isn't really needed. Will be removed in 1.0")
     override val serializationLoader: ServiceLoader<SerializationProvider> by lazy {
         val service = SerializationProvider::class.java
         ServiceLoader.load(service, service.classLoader)
@@ -58,22 +68,34 @@ public actual object XmlStreaming : XmlStreamingJavaCommon() {
         return factory.newWriter(result, repairNamespaces)
     }
 
-    override fun newWriter(outputStream: OutputStream, encoding: String, repairNamespaces: Boolean): XmlWriter {
+    public fun newWriter(outputStream: OutputStream, encoding: String, repairNamespaces: Boolean): XmlWriter {
         return factory.newWriter(outputStream, encoding, repairNamespaces)
     }
 
-    public actual override fun newWriter(writer: Writer, repairNamespaces: Boolean, xmlDeclMode: XmlDeclMode): XmlWriter {
-        return factory.newWriter(writer, repairNamespaces, xmlDeclMode)
+    @Deprecated("Use extension function on IXmlStreaming", level = DeprecationLevel.WARNING)
+    public actual fun newWriter(
+        writer: Writer,
+        repairNamespaces: Boolean /*= false*/,
+        xmlDeclMode: XmlDeclMode /*= XmlDeclMode.None*/,
+    ): XmlWriter {
+        return factory.newWriter(writer, repairNamespaces = repairNamespaces, xmlDeclMode = xmlDeclMode)
     }
 
-    public actual override fun newWriter(output: Appendable, repairNamespaces: Boolean, xmlDeclMode: XmlDeclMode): XmlWriter {
+    @Deprecated("Use extension function on IXmlStreaming", level = DeprecationLevel.WARNING)
+    public actual fun newWriter(
+        output: Appendable,
+        repairNamespaces: Boolean /*= false*/,
+        xmlDeclMode: XmlDeclMode /*= XmlDeclMode.None*/,
+    ): XmlWriter {
         return factory.newWriter(output, repairNamespaces, xmlDeclMode)
     }
 
+
+    @Deprecated("Use extension function on IXmlStreaming", level = DeprecationLevel.WARNING)
     public actual fun newGenericWriter(
         output: Appendable,
-        isRepairNamespaces: Boolean,
-        xmlDeclMode: XmlDeclMode
+        isRepairNamespaces: Boolean /*= false*/,
+        xmlDeclMode: XmlDeclMode /*= XmlDeclMode.None*/,
     ): KtXmlWriter {
         return KtXmlWriter(output, isRepairNamespaces, xmlDeclMode)
     }
@@ -82,7 +104,7 @@ public actual object XmlStreaming : XmlStreamingJavaCommon() {
         return factory.newReader(inputStream, encoding)
     }
 
-    actual override fun newReader(reader: Reader): XmlReader {
+    override fun newReader(reader: Reader): XmlReader {
         return factory.newReader(reader)
     }
 
@@ -90,15 +112,14 @@ public actual object XmlStreaming : XmlStreamingJavaCommon() {
         return factory.newReader(source)
     }
 
-    public actual override fun newReader(input: CharSequence): XmlReader {
+    public override fun newReader(input: CharSequence): XmlReader {
         return factory.newReader(input)
     }
 
-    override fun newReader(inputStr: String): XmlReader {
-        return factory.newReader(inputStr)
-    }
+    @Deprecated("Use the version taking a CharSequence")
+    public override fun newReader(input: String) : XmlReader = newReader(input as CharSequence)
 
-    public actual fun newGenericReader(input: CharSequence): XmlReader =
+    public override fun newGenericReader(input: CharSequence): XmlReader =
         newGenericReader(StringReader(input.toString()))
 
     public fun newGenericReader(input: String): XmlReader =
@@ -107,9 +128,9 @@ public actual object XmlStreaming : XmlStreamingJavaCommon() {
     public fun newGenericReader(inputStream: InputStream, encoding: String?): XmlReader =
         KtXmlReader(inputStream, encoding)
 
-    public actual fun newGenericReader(reader: Reader): XmlReader = KtXmlReader(reader)
+    public override fun newGenericReader(reader: Reader): XmlReader = KtXmlReader(reader)
 
-    public actual override fun setFactory(factory: XmlStreamingFactory?) {
+    public override fun setFactory(factory: XmlStreamingFactory?) {
         _factory = factory ?: AndroidStreamingFactory()
     }
 
@@ -147,3 +168,21 @@ public inline fun <reified T : Any> deserialize(input: InputStream): T = deSeria
 public inline fun <reified T : Any> deserialize(input: Reader): T = deSerialize(input, T::class.java)
 
 public inline fun <reified T : Any> deserialize(input: String): T = deSerialize(input, T::class.java)
+
+@Suppress("DEPRECATION")
+public actual val xmlStreaming: IXmlStreaming
+    get() = XmlStreaming
+
+@Suppress("DEPRECATION")
+public actual fun IXmlStreaming.newWriter(
+    output: Appendable,
+    repairNamespaces: Boolean,
+    xmlDeclMode: XmlDeclMode
+): XmlWriter = XmlStreaming.newWriter(output, repairNamespaces, xmlDeclMode)
+
+@Suppress("DEPRECATION")
+public actual fun IXmlStreaming.newWriter(
+    writer: Writer,
+    repairNamespaces: Boolean,
+    xmlDeclMode: XmlDeclMode,
+): XmlWriter = XmlStreaming.newWriter(writer, repairNamespaces, xmlDeclMode)