+++ core/src/commonMain/kotlin/nl/adaptivity/xmlutil/XmlStreaming.kt
@@ -24,16 +24,9 @@ import nl.adaptivity.xmlutil.core.KtXmlWriter
 import nl.adaptivity.xmlutil.core.impl.multiplatform.Reader
 import nl.adaptivity.xmlutil.core.impl.multiplatform.Writer
 
-/**
- * This class is the entry point for creating [XmlReader] and [XmlWriter]
- * instances. Some interfaces are common, others are limited to some
- * architectures.
- */
-public expect object XmlStreaming {
+public interface IXmlStreaming {
 
     public fun setFactory(factory: XmlStreamingFactory?)
-
-    public inline fun <reified T : Any> deSerialize(input: String): T
 
     public fun newReader(input: CharSequence): XmlReader
 
@@ -42,33 +35,78 @@ public expect object XmlStreaming {
     public fun newGenericReader(input: CharSequence): XmlReader
 
     public fun newGenericReader(reader: Reader): XmlReader
+}
 
-    @Deprecated("Use the version that takes an xmlDeclMode")
-    public fun newWriter(output: Appendable, repairNamespaces: Boolean = false, omitXmlDecl: Boolean): XmlWriter
 
-    public fun newWriter(output: Appendable, repairNamespaces: Boolean = false, xmlDeclMode: XmlDeclMode = XmlDeclMode.None): XmlWriter
+public expect fun IXmlStreaming.newWriter(
+    output: Appendable,
+    repairNamespaces: Boolean = false,
+    xmlDeclMode: XmlDeclMode = XmlDeclMode.None
+): XmlWriter
 
-    public fun newGenericWriter(output: Appendable, isRepairNamespaces: Boolean = false, xmlDeclMode: XmlDeclMode = XmlDeclMode.None): KtXmlWriter
+public expect fun IXmlStreaming.newWriter(
+    writer: Writer,
+    repairNamespaces: Boolean = false,
+    xmlDeclMode: XmlDeclMode = XmlDeclMode.None
+): XmlWriter
 
-    @Deprecated("Use the version that takes an xmlDeclMode")
-    public fun newWriter(writer: Writer, repairNamespaces: Boolean = false, omitXmlDecl: Boolean): XmlWriter
 
-    public fun newWriter(writer: Writer, repairNamespaces: Boolean, xmlDeclMode: XmlDeclMode = XmlDeclMode.None): XmlWriter
+public fun IXmlStreaming.newGenericWriter(
+    output: Appendable,
+    isRepairNamespaces: Boolean = false,
+    xmlDeclMode: XmlDeclMode = XmlDeclMode.None
+): KtXmlWriter = KtXmlWriter(output, isRepairNamespaces, xmlDeclMode)
+
+
+public expect val xmlStreaming: IXmlStreaming
+
+/**
+ * This class is the entry point for creating [XmlReader] and [XmlWriter]
+ * instances. Some interfaces are common, others are limited to some
+ * architectures.
+ */
+@Deprecated("Don't use directly", ReplaceWith("xmlStreaming",
+    "nl.adaptivity.xmlutil.xmlStreaming",
+    "nl.adaptivity.xmlutil.newWriter",
+    "nl.adaptivity.xmlutil.newGenericWriter",
+))
+public expect object XmlStreaming : IXmlStreaming {
+    public fun newGenericWriter(
+        output: Appendable,
+        isRepairNamespaces: Boolean = false,
+        xmlDeclMode: XmlDeclMode = XmlDeclMode.None
+    ): KtXmlWriter
+
+    public fun newWriter(
+        output: Appendable,
+        repairNamespaces: Boolean = false,
+        xmlDeclMode: XmlDeclMode = XmlDeclMode.None
+    ): XmlWriter
+
+    public fun newWriter(
+        writer: Writer,
+        repairNamespaces: Boolean = false,
+        xmlDeclMode: XmlDeclMode = XmlDeclMode.None
+    ): XmlWriter
+
 }
 
 
 public enum class XmlDeclMode {
     /** Don't emit XML Declaration */
     None,
+
     /** Emit an xml declaration just containing the xml version number. Only charsets that aren't UTF will be emitted. */
     Minimal,
+
     /** Emit an xml declaration whatever is provided by default, if possible minimal. */
     Auto,
+
     /** Emit an xml declaration that includes the character set. */
     Charset;
 
     internal companion object {
-        fun from(value: Boolean): XmlDeclMode = when(value) {
+        fun from(value: Boolean): XmlDeclMode = when (value) {
             true -> None
             else -> Auto
         }