+++ core/src/jvmMain/kotlin/nl/adaptivity/xmlutil/JVMXmlStreaming.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2023.
+ *
+ * This file is part of xmlutil.
+ *
+ * This file is licenced to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You should have received a copy of the license with the source distribution.
+ * Alternatively, you may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+// needed for abi compatibility
+@file:JvmName("JVMXmlStreamingKt")
+
+package nl.adaptivity.xmlutil
+
+import java.io.InputStream
+import java.io.Reader
+
+
+@Suppress("DEPRECATION")
+@Deprecated("This functionality uses service loaders and isn't really needed. Will be removed in 1.0")
+public inline fun <reified T : Any> deserialize(input: InputStream): T = XmlStreaming.deSerialize(input, T::class.java)
+
+@Suppress("DEPRECATION")
+@Deprecated("This functionality uses service loaders and isn't really needed. Will be removed in 1.0")
+public inline fun <reified T : Any> deserialize(input: Reader): T = XmlStreaming.deSerialize(input, T::class.java)
+
+@Suppress("DEPRECATION")
+@Deprecated("This functionality uses service loaders and isn't really needed. Will be removed in 1.0")
+public inline fun <reified T : Any> deserialize(input: String): T = XmlStreaming.deSerialize(input, T::class.java)